refactor: extract identity realm workflow service
