+++ server/src/main/java/com/massimotter/weave/backend/config/TimeConfiguration.java
@@ -1,0 +1,14 @@
+package com.massimotter.weave.backend.config;
+
+import java.time.Clock;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class TimeConfiguration {
+
+    @Bean
+    Clock systemClock() {
+        return Clock.systemUTC();
+    }
+}
